fixed to work on windows

+++ srcs/porting/open_file2.c
@@ -23,7 +23,7 @@ t_texture	*get_texture(unsigned char *buf, unsigned int *pos, ssize_t size)
 	while (++i < 3)
 	{
 		(*pos) += get_next_breaker(buf + (*pos)) + 1;
-		if (*pos > size)
+		if (*pos > (unsigned int)size)
 			error_output("Pointer points outside memory address\n");
 		if (i == 0)
 			idx = ft_atoi((char *)(buf + (*pos)));
@@ -34,7 +34,7 @@ t_texture	*get_texture(unsigned char *buf, unsigned int *pos, ssize_t size)
 			setup_parser(&png, png.source.size);
 			ft_memcpy(png.source.buf, buf + *pos, png.source.size);
 			(*pos) += png.source.size;
-			if (*pos > size)
+			if (*pos > (unsigned int)size)
 				error_output("Pointer points outside memory address\n");
 		}
 	}
@@ -65,7 +65,7 @@ void	parse_texture_data(unsigned char *buf, t_home *home,
 	*pos = 0;
 	buf = (unsigned char *)ft_strstr((char *)buf, "doom_textures");
 	*pos += get_next_breaker(buf + *pos) + 1;
-	if (*pos > size)
+	if (*pos > (unsigned int)size)
 		error_output("Pointer points outside memory address\n");
 	home->nbr_of_textures = ft_atoi((char *)buf + *pos);
 	home->editor_tex = (t_texture **)malloc(sizeof(t_texture *)
@@ -87,18 +87,18 @@ void	parse_audio_data(unsigned char *buf, unsigned int *pos,
 	t_audio_asset	asset;
 
 	*pos += get_next_breaker(buf + *pos) + 1;
-	if (*pos > size)
+	if (*pos > (unsigned int)size)
 		error_output("Pointer points outside memory address\n");
 	asset.size = ft_atoi((char *)buf + *pos);
 	asset.buf = malloc(sizeof(unsigned char *) * asset.size);
 	if (!asset.buf)
 		error_output("failed to allocate memory to music buffer\n");
 	*pos += get_next_breaker(buf + *pos) + 1;
-	if (*pos > size)
+	if (*pos > (unsigned int)size)
 		error_output("Pointer points outside memory address\n");
 	ft_memcpy(asset.buf, buf + *pos, asset.size);
 	*pos += asset.size;
-	if (*pos > size)
+	if (*pos > (unsigned int)size)
 		error_output("Pointer points outside memory address\n");
 	if (create_temp_audio_file(asset.buf, asset.size, path) == -1)
 		error_output("Failed to create temp audio file\n");
